refactor: 이전 구현된 컴포넌트 이동
