+++ lib/auth/presentation/components/auth_header_widget.dart
