+++ lib/auth/presentation/sign_in/screen/sign_in_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 import 'package:mongo_ai/auth/presentation/sign_in/controller/sign_in_action.dart';
 import 'package:mongo_ai/auth/presentation/sign_in/controller/sign_in_state.dart';
-import 'package:mongo_ai/core/component/auth_header_widget.dart';
+import 'package:mongo_ai/auth/presentation/components/auth_header_widget.dart';
 import 'package:mongo_ai/core/style/app_color.dart';
 import 'package:mongo_ai/core/style/app_text_style.dart';
